% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{bayesControls}
\alias{bayesControls}
\title{Control parameters for ic_bayes}
\usage{
bayesControls(samples = 10001, useMLE_start = TRUE, burnIn = 2499,
  iterationsPerUpdate = 250, initSD = 0.1, updateChol = T,
  acceptRate = 0.44, thin = 5)
}
\arguments{
\item{samples}{Number of samples.}

\item{useMLE_start}{Should MLE used for starting point? Highly recommended}

\item{burnIn}{Number of samples discarded for burn in}

\item{iterationsPerUpdate}{Number of iterations between updates of proposal covariance matrix}

\item{initSD}{If \code{useMLE_start == FALSE}, initial standard deviation used}

\item{updateChol}{Should cholesky decomposition be updated?}

\item{acceptRate}{Target acceptance rate}

\item{thin}{Amount of thinning}
}
\description{
Control parameters for ic_bayes
}

