% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utilities.R
\name{ir_sample}
\alias{ir_sample}
\title{Draw samples from an icenReg model}
\usage{
ir_sample(fit, newdata = NULL, sampleType = "fullSample", samples = 5)
}
\arguments{
\item{fit}{icenReg model fit}

\item{newdata}{\code{data.frame} containing covariates. If blank, will use data from model}

\item{sampleType}{type of samples See details for options}

\item{samples}{Number of samples}
}
\description{
Samples response values from an icenReg fit conditional on covariates.
}
\details{
If \code{newdata} is left blank, will provide estimates for original data set. 

 There are several options for how to sample. To get random samples without accounting
 for error in the estimated parameters \code{imputeType ='fixedParSample'} takes a 
 random sample of the response variable, conditional on the response interval, 
 covariates and estimated parameters at the MLE. Alternatively, 
 \code{imputeType = 'fullSample'} first takes a random sample of the coefficients,
 (assuming asymptotic normality for the ic_par) and then takes a random sample 
 of the response variable, conditional on the response interval, 
 covariates, and the random sample of the coefficients.
}
\examples{
simdata <- simIC_weib(n = 500, b1 = .3, b2 = -.3,
                      inspections = 6, inspectLength = 1)

fit <- ic_par(cbind(l, u) ~ x1 + x2,
              data = simdata)

newdata = data.frame(x1 = c(0, 1), x2 = c(1,1))

sampleResponses <- ir_sample(fit, newdata = newdata, samples = 100)
}
\author{
Clifford Anderson-Bergman
}
