% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utilities.R
\name{sampleSurv}
\alias{sampleSurv}
\title{Sample fitted survival function}
\usage{
sampleSurv(fit, newdata, p = NULL, q = NULL, samples = 100)
}
\arguments{
\item{fit}{Either an ic_bayes or ic_par fit}

\item{newdata}{A data.frame with a single row of covariates}

\item{p}{A set of survival probabilities to sample corresponding time for}

\item{q}{A set of times to sample corresponding survival probability for}

\item{samples}{Number of samples to draw}
}
\description{
Sample fitted survival function
}
\author{
Clifford Anderson-Bergman
}
