% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{simEventTime}
\alias{simEventTime}
\title{Simulates Event Time from Survival Regression Model}
\usage{
simEventTime(linPred = 0, model = "ph", dist = qweibull,
  paramList = list(shape = 1, scale = 1))
}
\arguments{
\item{linPred}{A numeric vector of linear predictors for
simulated data}

\item{model}{Regression model type. Options are \code{"ph"}, 
\code{"po"} or \code{"aft"}}

\item{dist}{The baseline distrubtion \code{q} function}

\item{paramList}{A list of parameters to be passed to the baseline
distribution function}
}
\description{
Simulates event times from proportional hazards,
proportional odds and accelerated failure time models.
}

