% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterBootstrap.R
\name{simIC_cluster}
\alias{simIC_cluster}
\title{Simulates data with multiple observations per subject}
\usage{
simIC_cluster(nIDs = 50, nPerID = 5)
}
\arguments{
\item{nIDs}{Number of subjects}

\item{nPerID}{Number of observations per subject}
}
\description{
Simulates data in which each subject is observed several times. 
In this case, the covariance matrix should be updated with \code{ir_clustBoot}.
}
