% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{bayesControls}
\alias{bayesControls}
\title{Control parameters for ic_bayes}
\usage{
bayesControls(samples = 5000, useMLE_start = TRUE, burnIn = 2500,
  iterationsPerUpdate = 250, updateChol = T, acceptRate = 0.44)
}
\arguments{
\item{samples}{Number of samples.}

\item{useMLE_start}{Should MLE used for starting point? Highly recommended}

\item{burnIn}{Number of samples discarded for burn in}

\item{iterationsPerUpdate}{Number of iterations between updates of proposal covariance matrix}

\item{updateChol}{Should cholesky decomposition be udpated?}

\item{acceptRate}{Target acceptance rate}

\item{cholScale}{Scalar value to adjust choleski proposal covariance matrix}

\item{burnIn}{Number of burn in samples used}
}
\description{
Control parameters for ic_bayes
}

